remove repo contents
* Update Settings.cs

* Update ILogicSettings.cs

* Update Settings.cs

* Update ILogicSettings.cs

* Update CatchPokemonTask.cs

* Fix CSV creation to use operating system culture

OS culture is used both in number format and in line separators

* Retaining old moves and placing them into new List List Moves structure

* Release prepartions

* Bahasa Indonesia: update translations.

Updated strings for first setup.

* Removed * at beginning line 0, position 0 #5340

* Bahasa Indonesia: fix typos.

Missing `-kan` suffix, `angka` also should've in lowercase...

* Fix Building Issue

I see this in almost all builds im not sure if this will fix it.

Description:
`Found conflicts between different versions of the same dependent assembly. Please set the "AutoGenerateBindingRedirects" property to true in the project file. For more information, see http://go.microsoft.com/fwlink/?LinkId=294190.`
File name:
`..\..\Program Files (x86)\MSBuild\14.0\bin\Microsoft.Common.CurrentVersion.targets`
Line:
`1820`
Column:
`5`
Project:
`PoGo.NecroBot.CLI`

* Remove AutoGeneratedBindingRedirects

* Removed extra space

* Added AutoGeneratedBindingRedirects and to true

* fixed snipe bug of going to already expired pokemon location

* Update translation.tr.json

* Update translation.tr.json

* Update translation.tr.json

* FIX translation.ja-JP.json

FIX to avoid the display bug.
And replacing the JP and JP2 because there is a need to change the default language file.

* Update translation.tr.json

* Update translation.tr.json

* FIX translation.ja-JP2.json

And replacing the JP and JP2 because there is a need to change the default language file.
This is a file for when the bugs have been removed in the future .
Please do not currently implemented .

* Update translation.tr.json

* fixed

* Random Walking Speed To Pokestops

* Fixed removed code

* Have random wait times at pokestops

* rewrote Encryption

* fixed compile-error

* Submodule sync

* compile-fix - 001

* update submodule

* submodule change

* submodule

* Remove post-builds

* farewell <3

* Update README.md

* remove repo contents